// frontend/src/app/generated/model/modelsIPGroup.ts
/**
 * Homelab API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export interface ModelsIPGroup {
  /**
   * 数据指纹，用于缓存失效
   */
  checksum?: string;
  createdAt?: string;
  description?: string;
  /**
   * 池中条目总数
   */
  entryCount?: number;
  id?: string;
  name?: string;
  updatedAt?: string;
}